+++ ghcloneall.py
@@ -29,7 +29,7 @@ import requests_cache
 __author__ = 'Marius Gedminas <marius@gedmin.as>'
 __licence__ = 'MIT'
 __url__ = 'https://github.com/mgedmin/ghcloneall'
-__version__ = '1.10.2.dev0'
+__version__ = '1.11.0'
 
 
 CONFIG_FILE = '.ghcloneallrc'